+++ MaterialDesignExtensions/Controls/OpenDirectoryDialog.cs
@@ -80,29 +80,6 @@ namespace MaterialDesignExtensions.Controls
         /// Shows a new <see cref="OpenDirectoryDialog" />.
         /// </summary>
         /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with OpenDirectoryDialogArguments instead")]
-        public static async Task<OpenDirectoryDialogResult> ShowDialogAsync(string dialogHostName, double? width = null, double? height = null,
-            string currentDirectory = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            OpenDirectoryDialog dialog = InitDialog(width, height, currentDirectory, false, showHiddenFilesAndDirectories, showSystemFilesAndDirectories);
-
-            return await DialogHost.Show(dialog, dialogHostName, openedHandler, closingHandler) as OpenDirectoryDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="OpenDirectoryDialog" />.
-        /// </summary>
-        /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
         /// <param name="args">The arguments for the dialog initialization</param>
         /// <returns></returns>
         public static async Task<OpenDirectoryDialogResult> ShowDialogAsync(string dialogHostName, OpenDirectoryDialogArguments args)
@@ -117,29 +94,6 @@ namespace MaterialDesignExtensions.Controls
             );
 
             return await DialogHost.Show(dialog, dialogHostName, args.OpenedHandler, args.ClosingHandler) as OpenDirectoryDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="OpenDirectoryDialog" />.
-        /// </summary>
-        /// <param name="dialogHost">The <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with OpenDirectoryDialogArguments instead")]
-        public static async Task<OpenDirectoryDialogResult> ShowDialogAsync(DialogHost dialogHost, double? width = null, double? height = null,
-            string currentDirectory = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            OpenDirectoryDialog dialog = InitDialog(width, height, currentDirectory, false, showHiddenFilesAndDirectories, showSystemFilesAndDirectories);
-
-            return await dialogHost.ShowDialog(dialog, openedHandler, closingHandler) as OpenDirectoryDialogResult;
         }
 
         /// <summary>