+++ MaterialDesignExtensions/Controls/OpenFileDialog.cs
@@ -47,29 +47,6 @@ namespace MaterialDesignExtensions.Controls
         /// Shows a new <see cref="OpenFileDialog" />.
         /// </summary>
         /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with OpenFileDialogArguments instead")]
-        public static async Task<OpenFileDialogResult> ShowDialogAsync(string dialogHostName, double? width = null, double? height = null,
-            string currentDirectory = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            OpenFileDialog dialog = InitDialog(width, height, currentDirectory, null, -1, showHiddenFilesAndDirectories, showSystemFilesAndDirectories);
-
-            return await DialogHost.Show(dialog, dialogHostName, openedHandler, closingHandler) as OpenFileDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="OpenFileDialog" />.
-        /// </summary>
-        /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
         /// <param name="args">The arguments for the dialog initialization</param>
         /// <returns></returns>
         public static async Task<OpenFileDialogResult> ShowDialogAsync(string dialogHostName, OpenFileDialogArguments args)
@@ -85,29 +62,6 @@ namespace MaterialDesignExtensions.Controls
             );
 
             return await DialogHost.Show(dialog, dialogHostName, args.OpenedHandler, args.ClosingHandler) as OpenFileDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="OpenFileDialog" />.
-        /// </summary>
-        /// <param name="dialogHost">The <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with OpenFileDialogArguments instead")]
-        public static async Task<OpenFileDialogResult> ShowDialogAsync(DialogHost dialogHost, double? width = null, double? height = null,
-            string currentDirectory = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            OpenFileDialog dialog = InitDialog(width, height, currentDirectory, null, -1, showHiddenFilesAndDirectories, showSystemFilesAndDirectories);
-
-            return await dialogHost.ShowDialog(dialog, openedHandler, closingHandler) as OpenFileDialogResult;
         }
 
         /// <summary>