+++ MaterialDesignExtensions/Controls/SaveFileDialog.cs
@@ -97,30 +97,6 @@ namespace MaterialDesignExtensions.Controls
         /// Shows a new <see cref="SaveFileDialog" />.
         /// </summary>
         /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="filename">The name of the file without the full path (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with SaveFileDialogArguments instead")]
-        public static async Task<SaveFileDialogResult> ShowDialogAsync(string dialogHostName, double? width = null, double? height = null,
-            string currentDirectory = null, string filename = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            SaveFileDialog dialog = InitDialog(width, height, currentDirectory, filename, null, -1, false, showHiddenFilesAndDirectories, showSystemFilesAndDirectories, false);
-
-            return await DialogHost.Show(dialog, dialogHostName, openedHandler, closingHandler) as SaveFileDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="SaveFileDialog" />.
-        /// </summary>
-        /// <param name="dialogHostName">The name of the <see cref="DialogHost" /></param>
         /// <param name="args">The arguments for the dialog initialization</param>
         /// <returns></returns>
         public static async Task<SaveFileDialogResult> ShowDialogAsync(string dialogHostName, SaveFileDialogArguments args)
@@ -139,30 +115,6 @@ namespace MaterialDesignExtensions.Controls
             );
 
             return await DialogHost.Show(dialog, dialogHostName, args.OpenedHandler, args.ClosingHandler) as SaveFileDialogResult;
-        }
-
-        /// <summary>
-        /// Shows a new <see cref="SaveFileDialog" />.
-        /// </summary>
-        /// <param name="dialogHost">The <see cref="DialogHost" /></param>
-        /// <param name="width">The width of the dialog (optional)</param>
-        /// <param name="height">The heigth of the dialog (optional)</param>
-        /// <param name="currentDirectory">The current directory to show (optional)</param>
-        /// <param name="filename">The name of the file without the full path (optional)</param>
-        /// <param name="showHiddenFilesAndDirectories">Show or hide hidden files in the dialog (optional)</param>
-        /// <param name="showSystemFilesAndDirectories">Show or hide system files in the dialog (optional)</param>
-        /// <param name="openedHandler">Callback after openening the dialog (optional)</param>
-        /// <param name="closingHandler">Callback after closing the dialog (optional)</param>
-        /// <returns></returns>
-        [Obsolete("Use the overloaded method with SaveFileDialogArguments instead")]
-        public static async Task<SaveFileDialogResult> ShowDialogAsync(DialogHost dialogHost, double? width = null, double? height = null,
-            string currentDirectory = null, string filename = null,
-            bool showHiddenFilesAndDirectories = false, bool showSystemFilesAndDirectories = false,
-            DialogOpenedEventHandler openedHandler = null, DialogClosingEventHandler closingHandler = null)
-        {
-            SaveFileDialog dialog = InitDialog(width, height, currentDirectory, filename, null, -1, false, showHiddenFilesAndDirectories, showSystemFilesAndDirectories, false);
-
-            return await dialogHost.ShowDialog(dialog, openedHandler, closingHandler) as SaveFileDialogResult;
         }
 
         /// <summary>